==> notebooks/5 - Weather .ipynb <==
import sqlite3
import pandas as pd

conn = sqlite3.connect('FPA_FOD_20170508.sqlite')
df = pd.read_sql_query("SELECT * from Fires", conn)

# load elevation - to collect the weather data
elevation_df = pd.read_csv("elevation.csv")

df.head(2)

elevation_df.head(2)

elevation_df.drop("Unnamed: 0", axis=1, inplace=True)
elevation_df.head(1)

print(elevation_df.shape, df.shape) # validate both with same number of rows - V

# Collect weather

## code to collect

import datetime as dt
import pandas as pd
import numpy as np
from meteostat import Point, Daily

df.head(1)

elevation_df.head(1)

**create seprate DISCOVERY date df, and convert it to dates**

date_df = df.DISCOVERY_DATE
date_df.head(1)

date_df = pd.to_datetime(date_df - pd.Timestamp(0).to_julian_date(), unit='D')
date_df.head(1)

**create df of LATITUDE, LONGITUDE, ELEVATION**

lat_long_el_df = pd.DataFrame({'latitude': np.array(df.LATITUDE),
                                'longitude': np.array(df.LONGITUDE),
                                'elevation': np.array(elevation_df.ELEVATION)})
lat_long_el_df.head(1)

# Number of days before the fire discovery day to collect data start from it.
NUM_DAYS_BEFORE = 1

# Expected num of days in collected data
total_days = NUM_DAYS_BEFORE + 1  

**create the empty new df to fill in the weather data**

def init_Point_object(i):
    # Create Point object for the fire location
    cur_fire_location = lat_long_el_df.iloc[i, :]
    location = Point(cur_fire_location["latitude"],
                     cur_fire_location["longitude"],
                     int(cur_fire_location["elevation"]))

    location.radius = 1e6  # Maximum radius for nearby stations in meters
    location.max_count = 5  # Maximum number of stations used for interpolation
    return location

def update_weather_df(weather_df, i, data):
    # The average air temperature in °C
    weather_df.iloc[i, TAVG_START:TAVG_END_PLUS_ONE] = data['tavg']

    # The minimum air temperature in °C
    weather_df.iloc[i, TMIN_START:TMIN_END_PLUS_ONE] = data['tmin']

    # The maximum air temperature in °C
    weather_df.iloc[i, TMAX_START:TMAX_END_PLUS_ONE] = data['tmax']

    # The average wind speed in km/h
    weather_df.iloc[i, TWSPD_START:TWSPD_END_PLUS_ONE] = data['wspd']

    # The daily precipitation total in mm
    weather_df.iloc[i, PRCP_START:PRCP_END_PLUS_ONE] = data['prcp']

    # The average sea-level air pressure in hPa
    weather_df.iloc[i, PRES_START:PRES_END_PLUS_ONE] = data['pres']

# create the new weather dataframe columns.
cols = ["tavg 0 days before", "tavg 1 days before",
       "tmin 0 days before", "tmin 1 days before",
       "tmax 0 days before", "tmax 1 days before",
       "wspd 0 days before", "wspd 1 days before",
       "prcp 0 days before", "prcp 1 days before",
       "pres 0 days before", "pres 1 days before",]

# columns index range in the new weather df
TAVG_START = 0
TAVG_END_PLUS_ONE = 2
TMIN_START = 2
TMIN_END_PLUS_ONE = 4
TMAX_START = 4
TMAX_END_PLUS_ONE = 6
TWSPD_START = 6
TWSPD_END_PLUS_ONE = 8
PRCP_START = 8
PRCP_END_PLUS_ONE = 10
PRES_START = 10
PRES_END_PLUS_ONE = 12

# new df shape
num_of_weather_data_columns_per_fire = len(cols)
num_of_fires = lat_long_el_df.shape[0]

# creating the weather df
weather_df = pd.DataFrame(np.full((num_of_fires,num_of_weather_data_columns_per_fire), np.nan), columns=cols)

## example - 5 rows

for i in range(5):
    # Set time period
    cur_fire_discovery_day = date_df[i]
    date_NUM_DAYS_BEFORE = cur_fire_discovery_day - dt.timedelta(days=NUM_DAYS_BEFORE)

    # init the Point object fot the fire location
    location = init_Point_object(i)

    # Get daily data for current fire for the fire discovered day and
    # one day before
    data = Daily(location, date_NUM_DAYS_BEFORE,
                 cur_fire_discovery_day)

    data = data.fetch()
    if data.shape[0] == total_days:  # validate that got the all expected days 
                                     # in the collected weather data. if not, skip.
        update_weather_df(weather_df, i, data)

weather_df

# all rows
takes a lot of time

# creating the weather df
weather_df = pd.DataFrame(np.full((num_of_fires,num_of_weather_data_columns_per_fire), np.nan), columns=cols)

for i in range(num_of_fires):
    # Set time period
    cur_fire_discovery_day = date_df[i]
    date_NUM_DAYS_BEFORE = cur_fire_discovery_day - dt.timedelta(days=NUM_DAYS_BEFORE)

    # init the Point object fot the fire location
    location = init_Point_object(i)

    # Get daily data for current fire for the fire discovered day and
    # one day before
    data = Daily(location, date_NUM_DAYS_BEFORE,
                 cur_fire_discovery_day)

    data = data.fetch()
    if data.shape[0] == total_days:  # validate that got the all expected days 
                                     # in the collected weather data. if not, skip.
        update_weather_df(weather_df, i, data)

weather_df

weather_df.isnull().sum() / ff.shape[0]

weather_df.to_csv("weather_df.csv")

med = weather_df.median()
med

weather_df.fillna(med, inplace=True)

weather_df

# can see that in the last row pres 0/1 filled as needed.
