==> notebooks/6 - All preprocess + choose model.ipynb <==
import sqlite3
import pandas as pd
import numpy as np

conn = sqlite3.connect('FPA_FOD_20170508.sqlite')
df = pd.read_sql_query("SELECT * from Fires", conn)

# selected features from 'First irrelevant features selection.ipynb'
# also no need for 'STATE' feature - already in 'LATITUDE','LONGITUDE'

features = ['CONT_DATE', 'CONT_DOY', 'CONT_TIME', 'DISCOVERY_DATE', 'DISCOVERY_DOY', 
            'DISCOVERY_TIME', 'FIRE_SIZE', 'FIRE_YEAR', 'LATITUDE','LONGITUDE', 
            'NWCG_REPORTING_AGENCY', 'OWNER_DESCR','STAT_CAUSE_CODE', 'STAT_CAUSE_DESCR']

### NWCG_REPORTING_AGENCY
will use as 5 dummies ST/C&L, FS, BIA, BLM and OTHER.

df.NWCG_REPORTING_AGENCY.unique()

NWCG_REPORTING_AGENCY = df.NWCG_REPORTING_AGENCY.apply(lambda x: x if x in ["ST/C&L", "FS", "BIA", "BLM"] else "Other")
NWCG_REPORTING_AGENCY.value_counts()

df.NWCG_REPORTING_AGENCY = NWCG_REPORTING_AGENCY

## OWNER_DESCR
we will use 4 dummies: MISSING/NOT SPECIFIED, PRIVATE, USFS and OTHER

df.OWNER_DESCR.unique()

OWNER_DESCR = df.OWNER_DESCR.apply(lambda x: x if x in ["MISSING/NOT SPECIFIED", "PRIVATE", "USFS"] else "Other")
OWNER_DESCR.value_counts()

df.OWNER_DESCR = OWNER_DESCR

# Time features

df['DISCOVERY_DATE'] = pd.to_datetime(df['DISCOVERY_DATE'] - pd.Timestamp(0).to_julian_date(), unit='D')
df['CONT_DATE'] = pd.to_datetime(df['CONT_DATE'] - pd.Timestamp(0).to_julian_date(), unit='D')

df['DISCOVERY_MONTH'] = df['DISCOVERY_DATE'].dt.month
df['DISCOVERY_DAY'] = df['DISCOVERY_DATE'].dt.weekday

df['DISCOVERY_TIME'] = pd.to_datetime(df['DISCOVERY_TIME'], format='%H%M')
df['CONT_TIME'] = pd.to_datetime(df['CONT_TIME'], format='%H%M', errors='coerce')

## burn days

df['BURN_DAYS'] = (df['CONT_DATE'] - df['DISCOVERY_DATE']).astype('timedelta64[D]')

df['DISCOVERY_DATE'].head(2)

df['CONT_DATE'].head(2)

df['BURN_DAYS'].head(2) # will fill nulls according to train median

## cyclic time

DAYS_IN_YEAR = 365
DAYS_IN_WEEK = 7

# no need for month cyclic - the data can be derived from DOY (day of year).

df['SIN_DISC_DOY'] = np.sin(2*np.pi*df['DISCOVERY_DOY']/DAYS_IN_YEAR)
df['COS_DISC_DOY'] = np.cos(2*np.pi*df['DISCOVERY_DOY']/DAYS_IN_YEAR)

df['SIN_DISC_WEEKDAY'] = np.sin(2*np.pi*df['DISCOVERY_DAY']/DAYS_IN_WEEK)
df['COS_DISC_WEEKDAY'] = np.cos(2*np.pi*df['DISCOVERY_DAY']/DAYS_IN_WEEK)

## weekend

df['IS_WEEKEND'] = (df['DISCOVERY_DAY'].isin([5, 6]))

## Independence day

df['INDEPENDENCE_DAY'] = ((df['DISCOVERY_DATE'].dt.day == 4) & (df['DISCOVERY_MONTH'] == 7))

# Elevation

elevation = pd.read_csv("elevation.csv")
elevation.head(1)

elevation.ELEVATION.isnull().sum() # no null

elevation.drop("Unnamed: 0", axis=1, inplace=True)
elevation.head(1)

# add elevation data to df
df = pd.concat([df, elevation], axis=1)
df

# Weather 

weather_df = pd.read_csv("weather_df.csv")
weather_df.head(1)

weather_df.drop("Unnamed: 0", axis=1, inplace=True)
weather_df.head(1)

weather_df.isnull().sum() / weather_df.shape[0] # will fill nulls with train median

# add weather_df data to df
df = pd.concat([df, weather_df], axis=1)
df

not relevant to preprocess: 'FIRE_SIZE', 'FIRE_YEAR', 'LATITUDE','LONGITUDE', 'STAT_CAUSE_CODE', 'STAT_CAUSE_DESCR'

# drop duplicates

df.shape

# remove duplicates observations: same location, date and  STAT_CAUSE_CODE
no_dups = df.drop_duplicates(subset=["DISCOVERY_DATE", "LATITUDE", "LONGITUDE", "STAT_CAUSE_CODE"])

no_dups.shape

# remove inconsistent observations: same location and  date but diffrent STAT_CAUSE_CODE
dup_check = no_dups[["DISCOVERY_DATE", "LATITUDE", "LONGITUDE", "STAT_CAUSE_CODE"]]

dup_check[dup_check[["DISCOVERY_DATE", "LATITUDE", "LONGITUDE"]].duplicated(keep=False)] # False : Mark all duplicates as True.

# we can see 1026, 1030 both has same location and date but deffrent cause - remove them.
no_dups = no_dups[~dup_check[["DISCOVERY_DATE", "LATITUDE", "LONGITUDE"]].duplicated(keep=False)]
no_dups.shape 

reduced from 1880465 to 1847186 <br>
run on the df

# remove duplicates observations: same location, date and  STAT_CAUSE_CODE
# df = df.drop_duplicates(subset=["DISCOVERY_DATE", "LATITUDE", "LONGITUDE", "STAT_CAUSE_CODE"])
dup_check = df[["DISCOVERY_DATE", "LATITUDE", "LONGITUDE", "STAT_CAUSE_CODE"]]
# remove inconsistent observations: same location and  date but diffrent STAT_CAUSE_CODE
df = df[~dup_check[["DISCOVERY_DATE", "LATITUDE", "LONGITUDE"]].duplicated(keep=False)]

df.shape

# end of preprocess - save preprocessed df, to skip it next time
df.to_csv("preprocessed.csv")

# train_test_split_df function

from sklearn.model_selection import train_test_split

def train_test_split_df(df):
    target = "STAT_CAUSE_CODE"
    
    # too many nulls in burn days, to remove? 
    time_features = ["FIRE_YEAR", "SIN_DISC_DOY", "COS_DISC_DOY", "SIN_DISC_WEEKDAY",  
                     "COS_DISC_WEEKDAY", "IS_WEEKEND", "INDEPENDENCE_DAY", "BURN_DAYS"]
   
    weather_features = ["tavg 0 days before", "tavg 1 days before", "tmin 0 days before", "tmin 1 days before",
                        "tmax 0 days before", "tmax 1 days before", "wspd 0 days before", "wspd 1 days before", 
                        "prcp 0 days before", "prcp 1 days before", "pres 0 days before", "pres 1 days before" ]
   
    geographical_features = ['LATITUDE', 'LONGITUDE', 'ELEVATION']
    
    other_features = ["FIRE_SIZE", "NWCG_REPORTING_AGENCY", 'OWNER_DESCR']
    
    y = df[target]
    X = df[time_features + weather_features + geographical_features + other_features]
    
    # split with stratify - for the unbalanced target.
    X_train, X_test, y_train, y_test = train_test_split(X, y,test_size=0.25, stratify=y)
    
    # get dummies 
    categorical_cols = ["NWCG_REPORTING_AGENCY", 'OWNER_DESCR']
    
    X_train = pd.get_dummies(X_train, columns=categorical_cols, drop_first=True) 
    X_test = pd.get_dummies(X_test, columns=categorical_cols, drop_first=True) 
    
    # match the dummies
    cols_in_both = sorted(set(X_train.columns).intersection(set(X_test.columns)))
    X_train = X_train[cols_in_both]
    X_test = X_test[cols_in_both]
    
    return X_train, X_test, y_train, y_test

X_train, X_test, y_train, y_test = train_test_split_df(df)

X_train.isnull().sum() / X_train.shape[0]

X_test.isnull().sum() / X_test.shape[0]

def fill_na_Train_Test(X_train, X_test):
    X_train_median = X_train.median()
    X_train = X_train.fillna(X_train_median)
    X_test = X_test.fillna(X_train_median)
    return X_train, X_test

X_train.median()

X_train, X_test = fill_na_Train_Test(X_train, X_test)

X_train.isnull().sum() 

X_test.isnull().sum() 

**no nulls**

# Choose model
failed with lazy predict, check between multiple which is the best, after choosing model will do hyper-parameters tuning for him.

def get_X_y(df):
    target = "STAT_CAUSE_CODE"

    # too many nulls in burn days, to remove?
    time_features = ["FIRE_YEAR", "SIN_DISC_DOY", "COS_DISC_DOY", "SIN_DISC_WEEKDAY", 
                     "COS_DISC_WEEKDAY", "IS_WEEKEND", "INDEPENDENCE_DAY", "BURN_DAYS"]

    weather_features = ["tavg 0 days before", "tavg 1 days before", "tmin 0 days before", "tmin 1 days before",
                        "tmax 0 days before", "tmax 1 days before", "wspd 0 days before", "wspd 1 days before", 
                        "prcp 0 days before", "prcp 1 days before", "pres 0 days before", "pres 1 days before" ]

    geographical_features = ['LATITUDE', 'LONGITUDE', 'ELEVATION']

    other_features = ["FIRE_SIZE", "NWCG_REPORTING_AGENCY", 'OWNER_DESCR']
    
    y = df[target]
    X = df[time_features + weather_features + geographical_features + other_features]
    # get dummies 
    categorical_cols = ["NWCG_REPORTING_AGENCY", 'OWNER_DESCR']
    X = pd.get_dummies(X, columns=categorical_cols, drop_first=True) 
    
    return X, y

target = ["STAT_CAUSE_CODE"]

time_features = ["FIRE_YEAR", "SIN_DISC_DOY", "COS_DISC_DOY", "SIN_DISC_WEEKDAY", 
                 "COS_DISC_WEEKDAY", "IS_WEEKEND", "INDEPENDENCE_DAY", "BURN_DAYS"]

weather_features = ["tavg 0 days before", "tavg 1 days before", "tmin 0 days before", "tmin 1 days before",
                    "tmax 0 days before", "tmax 1 days before", "wspd 0 days before", "wspd 1 days before", 
                    "prcp 0 days before", "prcp 1 days before", "pres 0 days before", "pres 1 days before" ]

geographical_features = ['LATITUDE', 'LONGITUDE', 'ELEVATION']

other_features = ["FIRE_SIZE", "NWCG_REPORTING_AGENCY", 'OWNER_DESCR']
features = time_features + weather_features + geographical_features + other_features + target

df = pd.read_csv("preprocessed.csv")[features]

df.head(2)

# drop null values before cross_val_score 
df_no_null = df.dropna()
df_no_null

# sample stratified subset of the data to reduce cross_val_score running time.

# https://www.statology.org/stratified-sampling-pandas/
N = int(0.3 * df_no_null.shape[0])
samp = df_no_null.groupby('STAT_CAUSE_CODE', group_keys=False).apply(lambda x: x.sample(int(np.rint(N*len(x)/len(df_no_null))))).sample(frac=1).reset_index(drop=True)
samp

# Cross-Val on classifiers to pick the best between them.
Used random_state=1 to allow the results to be restored.

# import sklearn
#sklearn.metrics.SCORERS.keys()

**gradientBoostingClassifier:**

# https://machinelearningmastery.com/gradient-boosting-with-scikit-learn-xgboost-lightgbm-and-catboost/
## gradientBoostingClassifier
from numpy import mean
from numpy import std
from sklearn.datasets import make_classification
from sklearn.ensemble import GradientBoostingClassifier
from sklearn.model_selection import cross_val_score
from sklearn.model_selection import RepeatedStratifiedKFold

X, y = get_X_y(samp)

model = GradientBoostingClassifier(random_state=1)
cv = RepeatedStratifiedKFold(n_splits=5, n_repeats=3, random_state=1)

n_scores = cross_val_score(model, X, y, scoring='f1_weighted', cv=cv, n_jobs=-1, error_score='raise')

n_scores

print('f1_weighted: mean=%.3f std=%.3f' % (mean(n_scores), std(n_scores)))

**XGBClassifier:**

from xgboost import XGBClassifier

X, y = get_X_y(samp)
y = y - 1 # for 'num_class' value in cv params
  
model = XGBClassifier(random_state=1)

cv = RepeatedStratifiedKFold(n_splits=5, n_repeats=3, random_state=1)

n_scores = cross_val_score(model, X, y, scoring='f1_weighted', cv=cv, n_jobs=-1, error_score='raise')

n_scores

print('f1_weighted: mean=%.3f std=%.3f' % (mean(n_scores), std(n_scores)))

**RFClassifier:**

## RFClassifier

# https://machinelearningmastery.com/random-forest-ensemble-in-python/
# evaluate random forest algorithm for classification
from numpy import mean
from numpy import std
from sklearn.datasets import make_classification
from sklearn.model_selection import cross_val_score
from sklearn.model_selection import RepeatedStratifiedKFold
from sklearn.ensemble import RandomForestClassifier
# define dataset
X, y =  get_X_y(samp)
X.fillna(X.median(), inplace=True)
# define the model
model = RandomForestClassifier(random_state=1)
# evaluate the model
cv = RepeatedStratifiedKFold(n_splits=5, n_repeats=3, random_state=1)

n_scores = cross_val_score(model, X, y, scoring='f1_weighted', cv=cv, n_jobs=-1, error_score='raise')

n_scores

# report performance
print('f1_weighted: mean=%.3f std=%.3f' % (mean(n_scores), std(n_scores)))

# Hyperparameter tuning 
RandomizedSearchCV

# https://jamesrledoux.com/code/randomized_parameter_search

from scipy import stats
from scipy.stats import randint
from sklearn.model_selection import RandomizedSearchCV
from xgboost import XGBClassifier

N = int(0.1 * df_no_null.shape[0])
samp = df_no_null.groupby('STAT_CAUSE_CODE', group_keys=False).apply(lambda x: x.sample(int(np.rint(N*len(x)/len(df_no_null))))).sample(frac=1).reset_index(drop=True)
samp

X, y = get_X_y(samp)
y = y - 1 # for 'num_class' value in cv params

len(y.unique())

#https://stackoverflow.com/questions/43927725/python-hyperparameter-optimization-for-xgbclassifier-using-randomizedsearchcv

# multi:softmax: set XGBoost to do multiclass classification using the 
# softmax objective, you also need to set num_class(number of classes)

# Changed to: objective = 'multi:softmax'.
clf_xgb = XGBClassifier(objective = 'multi:softmax', num_class=len(y.unique()))

param_dist = {'n_estimators': stats.randint(150, 1000),
              'learning_rate': stats.uniform(0.01, 0.59),
              'subsample': stats.uniform(0.3, 0.6),
              'max_depth': [3, 4, 5, 6, 7, 8, 9],
              'colsample_bytree': stats.uniform(0.5, 0.4),
              'min_child_weight': [1, 2, 3, 4]
             }

numFolds = 5
kfold_5 = RepeatedStratifiedKFold(n_splits=5, n_repeats=3, random_state=1)

clf = RandomizedSearchCV(clf_xgb, 
                         param_distributions = param_dist,
                         cv = kfold_5,  
                         n_iter = 5, 
                         scoring = 'f1_weighted', 
                         error_score = 0, 
                         verbose = 1)

clf.fit(X, y)

clf.best_params_

clf.best_score_

# Train-Test on all data

from xgboost import XGBClassifier
from sklearn.model_selection import train_test_split
import pandas as pd
import numpy as np

df = pd.read_csv("preprocessed.csv")

def train_test_split_df(df):
    target = "STAT_CAUSE_CODE"
    
    time_features = ["FIRE_YEAR", "SIN_DISC_DOY", "COS_DISC_DOY", "SIN_DISC_WEEKDAY", 
                     "COS_DISC_WEEKDAY", "IS_WEEKEND", "INDEPENDENCE_DAY", "BURN_DAYS"]
   
    weather_features = ["tavg 0 days before", "tavg 1 days before", "tmin 0 days before", "tmin 1 days before",
                        "tmax 0 days before", "tmax 1 days before", "wspd 0 days before", "wspd 1 days before", 
                        "prcp 0 days before", "prcp 1 days before", "pres 0 days before", "pres 1 days before" ]
   
    geographical_features = ['LATITUDE', 'LONGITUDE', 'ELEVATION']
    
    other_features = ["FIRE_SIZE", "NWCG_REPORTING_AGENCY", 'OWNER_DESCR']
    
    y = df[target]
    X = df[time_features + weather_features + geographical_features + other_features]
    
    # split with stratify - becaues of the unbalanced target.
    X_train, X_test, y_train, y_test = train_test_split(X, y,test_size=0.25, stratify=y)
    
    # get dummies 
    categorical_cols = ["NWCG_REPORTING_AGENCY", 'OWNER_DESCR']
    
    X_train = pd.get_dummies(X_train, columns=categorical_cols, drop_first=True) 
    X_test = pd.get_dummies(X_test, columns=categorical_cols, drop_first=True) 
    
    # match the dummies
    cols_in_both = sorted(set(X_train.columns).intersection(set(X_test.columns)))
    X_train = X_train[cols_in_both]
    X_test = X_test[cols_in_both]
    
    return X_train, X_test, y_train, y_test

def fill_na_Train_Test(X_train, X_test):
    X_train_median = X_train.median()
    X_train = X_train.fillna(X_train_median)
    X_test = X_test.fillna(X_train_median)
    return X_train, X_test

X_train, X_test, y_train, y_test = train_test_split_df(df)

y_train = y_train - 1
y_test = y_test - 1

X_train, X_test = fill_na_Train_Test(X_train, X_test)

XGB_best_params = {  'colsample_bytree': 0.5787,
                     'learning_rate': 0.0767,
                     'max_depth': 4,
                     'min_child_weight': 3,
                     'n_estimators': 839,
                     'subsample': 0.6643}

clf_xgb = XGBClassifier(objective = 'multi:softmax', num_class=13, **XGB_best_params)

clf_xgb.fit(X_train, y_train)

pred = clf_xgb.predict(X_test)

from sklearn.metrics import f1_score

f1_score(y_test, pred, average=None)

f1_score(y_test, pred, average='weighted')
